use correct seq name and remove abstract interface
