--- authorizer-app-backend/src/main/java/org/radarbase/authorizer/doa/AbstractJpaPersistable.kt
@@ -1,44 +1,0 @@
-/*
- *
- *  * Copyright 2019 The Hyve
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *   http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *  *
- *
- */
-
-package org.radarbase.authorizer.doa
-
-import javax.persistence.*
-
-@MappedSuperclass
-abstract class AbstractJpaPersistable<T : java.io.Serializable> {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
-    @SequenceGenerator(name = "sequenceGenerator", initialValue = 1000)
-    var id: T? = null
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-
-        if (other == null || javaClass != other.javaClass) return false
-
-        other as AbstractJpaPersistable<*>
-
-        return id != null && id == other.id
-    }
-
-    override fun hashCode(): Int = id.hashCode()
-
-    override fun toString() = "Entity of type ${this.javaClass.name} with id: $id"
-}
